Croeseid testnet getting-started: update create-validator transaction
For me and other members of the community on Discord 
(https://discord.com/channels/783264383978569728/783264902613565450/789208334800388127), 
we needed to add `gas-prices` to fix the `insufficient fee` error. 
Also, since we get 10 test CRO per day on the faucet,  I reduced the amount 
of the `create-validator` transaction, since a low amount worked for me.